Promenjeno formatiranje.

+++ strukture_B.cpp
@@ -1,6 +1,7 @@
 #include <iostream>
 #include <string>
 #include <deque>
+#include <cstdint>
 
 bool check(char a, char b)
 {
@@ -16,7 +17,7 @@ int main()
     std::string s;
     std::cin >> s;
     std::deque<char> d;
-    for (int32_t i = 0; i < s.length(); i++)
+    for (size_t i = 0; i < s.length(); i++)
     {
         if (d.empty())
         {
